requirements.txt
groq
fastapi
uvicorn
requests
python-dotenv
pandas